Add Trivia minigame (#34)
Adds a Trivia minigame, accessible through Trivia, with rounds that are won based on how many questions are correctly answered out of a set of N questions. Uses the Newtonsoft.json framework. Because Trivia reads its questions from a json file, creating an instance requires that the creator knows the path to the questions source file. This commit comes with 60 questions generated by ChatGPT, stored in Questions.json